moved getLockedBlock() into PersistentData.java
